Split MPSC endpoint implementation
Move sender and receiver code into focused modules. Keep public exports unchanged and make iterator construction const where possible.

Report endpoint disconnection when all senders are gone while retaining an internal drained check for terminal receive results.
